refactor(mentor): usePageSize.js 이름 useDeviceType.js로 변경
