+++ src/common/useDeviceType.js
